Move handler to be under server package
