+++ .golangci.yml
@@ -34,11 +34,12 @@ linters:
     - staticcheck # 30+files need to be patched
 issues:
   exclude-use-default: false
-  exclude-rules:
-    - path: cmd/easi/(serve|test)\.go
-      text: "should have a package comment, unless it's in another file for this package"
-      linters:
-        - golint
+
+  exclude:
+    #TODO: remove this exclusion after https://github.com/golang/lint/issues/381
+    # is fixed by https://github.com/golang/lint/pull/380 or something else
+    - "should have a package comment, unless it's in another file for this package"
+
 run:
   # timeout for analysis, e.g. 30s, 5m, default is 1m
   deadline: 8m
