+++ pkg/server/routes.go
@@ -1,0 +1,5 @@
+package server
+
+func (s *server) routes() {
+	s.router.HandleFunc("/", s.corsMiddleware(s.authorizeHandler(s.handleLanding())))
+}
