+++ pkg/server/server.go
@@ -1,4 +1,4 @@
-// Package server is some throwaway server code to setup testing
+// Package server is for setting up the server.
 package server
 
 import (
@@ -7,26 +7,19 @@ import (
 	"net/http"
 
 	"github.com/gorilla/mux"
-
-	"github.com/spf13/viper"
-
-	"github.com/cmsgov/easi-app/pkg/handlers"
 )
 
-func corsHandler(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-
-		w.Header().Set("Access-Control-Allow-Origin", viper.GetString("CLIENT_ADDRESS"))
-		w.Header().Set("Access-Control-Allow-Headers", "Accept, Content-Type, Content-Length, Accept-Encoding, Authorization")
-
-		next.ServeHTTP(w, r)
-	})
+type server struct {
+	router *mux.Router
 }
 
 // Serve serves all the handlers
 func Serve() {
 	r := mux.NewRouter()
+	s := server{
+		router: r,
+	}
 	fmt.Print("Serving application on localhost:8080")
-	r.Handle("/", corsHandler(authorizeHandler(handlers.LandingHandler{})))
-	log.Fatal(http.ListenAndServe(":8080", r))
+	s.routes()
+	log.Fatal(http.ListenAndServe(":8080", s.router))
 }
